+++ ClozeMaker.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClozeMaker", "ClozeMaker.csproj", "{0850B93E-6DCF-409B-BE45-1DD61900DB70}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClozeMaker", "src\ClozeMaker\ClozeMaker.csproj", "{0850B93E-6DCF-409B-BE45-1DD61900DB70}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
